Úprava vue

+++ WannabeGCalendar_ASP/WannabeGCalendar/WannabeGCalendar/Entities/Event.cs
@@ -10,7 +10,7 @@ namespace WannabeGCalendar.Entities
     {
         [Key]
         [Column("EventID")]
-        public Int64 EventId { get; set; }
+        public int EventId { get; set; }
         public string EventName { get; set; }
         public DateTime EventDate { get; set; }
         public TimeSpan? EventTime { get; set; }
